File Change:  Code Rewrite
Complete Rewrite of the code in a new file. It is now a lot easier to understand and edit as it is separated into more functions. Added lidar altitude control and adjusted the pos hold code to get the drone to fly between to target positions. Also added payload drop functionality.